requirements.txt
anyio==3.6.2
certifi==2022.9.24
click==8.1.3
dnspython==2.3.0
fastapi==0.90.0
h11==0.14.0
idna==3.4
loguru==0.6.0
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
sniffio==1.3.0
starlette==0.23.0
typing_extensions==4.4.0
uvicorn==0.20.0
